16530 - Fix plasma-js-bridge
The presence of an npm-shrinkwrap file that specified
`plasma-js-bridge@0.1.6` was causing this package to be built with the
incorrect version of the library than what was specified in
`package.json` (`plasma-js-bridge@0.1.8`).

Problem is, 0.1.8 contained a bug that was causing `plasma-web-proxy` to
crash on receiving a protein. It has been patched to version 0.1.10
which fixes this issue, so `package.json` has been updated, and
`npm-shrinkwrap.json` removed as it is no longer necessary.