16592 - Fix crash on protein without descrips|ingests
Plasma-web-proxy was crashing when it processed a protein without
descrips or ingests. This problem arose due to changes that made
`process_protein` call the provided callback with `null` for error
conditions, including no ingests or descrips. The fix is to augment
`zipProtein` to be tolerant of proteins without descrips or ingests.