19764 pause for hose setup
This adds a short buffer period between a connection being opened and
when it is reported to a client as being opened. This is to fix a potential
race condition of where the WebSocket client could start sending and
receiving messages before the peek hose was fully open and able to read
proteins.

The fix is a hack. From experimentation 10ms is ample time for
a peek process to fully set up for local pools, which is what the proxy
cares about. Still, it would be much better to have an explicit way to wait
for hose open, it's just not easy to do so with the given API.