Add ability for client to specify max retry attempts
This patch is mainly for screencast at the moment but could
come in handy for other clients as well. It allows a
client to specify a maximum number of retries on connection
attempts before giving up. This patch also allows the client
to choose to manually close the socket connection via the
closeSock method in Transport. Previously it was assumed that
the connection would close via the user closing their browser, but
in screencast we need a way to close the socket connection without
closing the browser (since the browser is technically open all the
time).

BZ-19772